Change cx/cy in azimuthal integration from int to float

+++ extra_foam/gui/mediator.py
@@ -149,10 +149,10 @@ class Mediator(QObject):
     def onAiPixelSizeYChange(self, value: int):
         self._meta.hset(mt.AZIMUTHAL_INTEG_PROC, 'pixel_size_y', value)
 
-    def onAiIntegCenterXChange(self, value: int):
+    def onAiIntegCenterXChange(self, value: float):
         self._meta.hset(mt.AZIMUTHAL_INTEG_PROC, 'integ_center_x', value)
 
-    def onAiIntegCenterYChange(self, value: int):
+    def onAiIntegCenterYChange(self, value: float):
         self._meta.hset(mt.AZIMUTHAL_INTEG_PROC, 'integ_center_y', value)
 
     def onAiIntegMethodChange(self, value: str):
